requirements.txt
# Core Data Science Libraries (Updated for Python 3.12+)
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Machine Learning
xgboost>=2.0.0
lightgbm>=4.1.0

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Model Explainability
shap>=0.44.0
lime>=0.2.0.1

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Database
sqlalchemy>=2.0.23

# API Integration
requests>=2.31.0
googlemaps>=4.10.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
joblib>=1.3.2

# Testing
pytest>=7.4.3

# Jupyter
jupyter>=1.0.0
ipykernel>=6.27.0
notebook>=7.0.6

# Progress bars
tqdm>=4.66.0

# Additional dependencies for better compatibility
setuptools>=69.0.0
wheel>=0.42.0